requirements.txt
uvicorn==0.23.2
fastapi==0.109.2
pandas==2.2.3
datetime==5.5
SQLAlchemy==2.0.24
google-cloud-secret-manager==2.19.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
asyncio==3.4.3
nest-asyncio==1.5.6
pg8000==1.29.8
cloud-sql-python-connector==1.2.4
boto3==1.34.125